Add files via upload
The original P/ESDM was created in Indonesian. In this English version, we translated the Indonesian variables into English, but it resulted in errors. Therefore, we recommend using the Indonesian version of P/ESDM, while referring to this English version for the meaning of the variables. We would greatly appreciate any assistance in improving this English version of P/ESDM.